# ci(perf/net_latency): increase absolute delta threshold to 10us